Feature/compras ventas

+++ Dominio/Productos.cs
@@ -12,7 +12,7 @@ namespace Dominio
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public decimal UltPrecio { get; set; }
-        public int PorcGanancia { get; set; }
+        public double PorcGanancia { get; set; }
         public string UrlImagen { get; set; }
         public Marca Marca { get; set; }
         public Categoria Categoria { get; set; }
